changed TCVS320 firmware  pin assignement
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@433 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/dollbrain/ColorEvent.java
@@ -17,13 +17,13 @@ import ch.unizh.ini.caviar.event.*;
  */
 public class ColorEvent extends TypedEvent {
     
-    public byte color=0;
+    public short color=0;
     
     /** Creates a new instance of TypedEvent */
     public ColorEvent() {
     }
     
-    public byte getColor(){
+    public short getColor(){
         return color;
     }
